pyproject.toml
[project]
name = "bronze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord>=2.3.2",
    "icecream>=2.1.4",
    "ipdb>=0.13.13",
    "ipython>=8.18.1",
    "litellm>=1.63.12",
    "load-dotenv>=0.1.0",
    "notion-client>=2.3.0",
    "pyreadline>=2.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "rich>=13.9.4",
    "textual>=2.1.2",
    "llmgine @ {root:parent:uri}/llmgine",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.86.0",
    "sqlalchemy>=2.0.40",
    "psycopg2-binary>=2.9.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bronze"]

[tool.hatch.metadata]
allow-direct-references = true
